requirements.txt
streamlit
pandas
pyyaml
simpleeval
